online-notes-sharing-python/requirements.txt
flask
pymongo
werkzeug
python-dotenv
flask-login
